fix: resolve CI lint failures for PR mergeability
Extract Hide/Show menu into its own component file to satisfy
noExcessiveLinesPerFile, prefix unused snapshot param, and inline
findNode lookup to trim factory-graph-editor-connections.
