+++ ui/src/features/factory-graph-editor/lib/factory-graph-editor-connections.ts
@@ -267,8 +267,10 @@ export function buildFactoryGraphEdgeChangeFromConnection(
   },
   resolver?: FactoryGraphConnectionResolver,
 ): FactoryGraphDraftEdgeChange | null {
-  const sourceNode = findNode(topology, endpoint.sourceNodeId);
-  const targetNode = findNode(topology, endpoint.targetNodeId);
+  const sourceNode =
+    topology.nodes.find((node) => node.id === endpoint.sourceNodeId) ?? null;
+  const targetNode =
+    topology.nodes.find((node) => node.id === endpoint.targetNodeId) ?? null;
   if (!sourceNode || !targetNode) {
     return null;
   }
@@ -480,6 +482,3 @@ function appendUniqueEdgeChange(
   return [...edges, edgeChange];
 }
 
-function findNode(topology: FactoryGraphTopology, nodeId: string) {
-  return topology.nodes.find((node) => node.id === nodeId) ?? null;
-}
